requirements.txt
accelerate==0.21.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
confection==0.1.0
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
decord==0.6.0
distlib==0.3.7
einops==0.6.1
executing==1.2.0
facenet-pytorch==2.5.3
fairscale==0.4.4
filelock==3.12.2
fonttools==4.41.1
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
identify==2.5.26
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
iopath==0.1.10
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kaggle==1.5.16
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nodeenv==1.8.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==1.5.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.9.1
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.0
regex==2022.9.13
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
salesforce-lavis==1.0.2
scikit-image==0.21.0
scipy==1.11.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
text-unidecode==1.3
thinc==8.1.10
tifffile==2023.7.18
timm==0.4.12
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
virtualenv==20.24.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
webencodings==0.5.1
zipp==3.16.2
